Banging my head on design and code
Thing is: I don't precisely know if the user will format by line.
Theoretically, pascal can be done just like C -> be a single line of big ass code and stuff, no newlines whatsoever. The bad way to do this is to assume the user will put them newlines on the program.
This ain't gonna cut it.
I'll try to come up with some way this kind of fuckup. Can't do this now though, so 'ill commit the changes and that's it.