Added most of the token finding stuff
next is the construction of them. Right now, the lexical analyzer should be able to count the tokens, but only that.
Plus, the solution implemented was very very dumb, but also, the remove method seems to simply remove the first occurrence of something, which ain't great. Had to resort to a try except inside a while twice to get rid of \n's and spaces, which would prolly Mike Foxtrot Uniform the analysis, not very great.